--- apps/ngx-env-demo/.dockerignore
@@ -1,4 +1,0 @@
-node_modules/
-.gitignore
-dist/
-.angular/
