+++ apps/ngx-env-demo/.env
@@ -1,3 +1,3 @@
 USER_HOME=/home/chihab
-NGX_VERSION=IGNORED_SINCE_SET_IN_ENV_LOCAL
-NGX_USER_HOME=$USER_HOME
+NG_APP_VERSION=IGNORED_SINCE_SET_IN_ENV_LOCAL
+NG_APP_USER_HOME=$USER_HOME
