+++ apps/ngx-env-demo/.env.local
@@ -1,1 +1,1 @@
-NGX_VERSION=$npm_package_version
+NG_APP_VERSION=17
