--- apps/ngx-env-demo/env.d.ts
@@ -1,17 +1,0 @@
-// This file should not be here. It is temporary until only import.meta.env is used.
-
-interface ImportMeta {
-  readonly env: ImportMetaEnv;
-}
-
-interface ImportMetaEnv {
-  /**
-   * Built-in environment variable.
-   * @see Docs https://github.com/chihab/ngx-env#ng_app_env.
-   */
-  readonly NG_APP_ENV: string;
-  // Add your environment variables below
-  // readonly NG_APP_API_URL: string;
-  [key: string]: any;
-}
-
