+++ apps/ngx-env-demo/package.json
@@ -1,54 +1,45 @@
 {
-  "name": "@ngx-env/demo",
-  "version": "16.1.1",
-  "private": true,
+  "name": "ngx-env-demo",
+  "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "start": "cross-env NODE_ENV=development ng serve",
-    "prepare:build": "npx ng config projects.ngx-env-demo.architect.build.options.ngxEnv.prefix 'NGX_'",
-    "build": "ng build",
-    "prepare:test": "npx ng config projects.ngx-env-demo.architect.test.options.ngxEnv.prefix 'NGX_'",
-    "test": "cross-env NODE_ENV=test NGX_BRANCH_NAME=main ng test --watch=false",
-    "prepare:server": "npx ng config projects.ngx-env-demo.architect.server.options.ngxEnv.prefix 'NGX_'",
-    "dev:ssr": "ng run ngx-env-demo:serve-ssr",
-    "serve:ssr": "node dist/ngx-env-demo/server/main.js",
-    "build:ssr": "ng build && ng run ngx-env-demo:server",
-    "prerender": "ng run ngx-env-demo:prerender"
+    "start": "NODE_ENV=development ng serve",
+    "build": "NODE_ENV=development ng build",
+    "watch": "ng build --watch --configuration development",
+    "test": "ng test"
   },
+  "private": true,
   "dependencies": {
-    "@angular/animations": "^16.0.0",
-    "@angular/common": "^16.0.0",
-    "@angular/compiler": "^16.0.0",
-    "@angular/core": "^16.0.0",
-    "@angular/forms": "^16.0.0",
-    "@angular/localize": "^16.0.0",
-    "@angular/platform-browser": "^16.0.0",
-    "@angular/platform-browser-dynamic": "^16.0.0",
-    "@angular/platform-server": "^16.0.0",
-    "@angular/router": "^16.0.0",
-    "@nguniversal/express-engine": "^16.0.0",
+    "@angular/animations": "^17.0.0",
+    "@angular/common": "^17.0.0",
+    "@angular/compiler": "^17.0.0",
+    "@angular/core": "^17.0.0",
+    "@angular/forms": "^17.0.0",
+    "@angular/platform-browser": "^17.0.0",
+    "@angular/platform-browser-dynamic": "^17.0.0",
+    "@angular/platform-server": "^17.0.0",
+    "@angular/router": "^17.0.0",
+    "@angular/ssr": "^17.0.0",
     "@ngx-env/core": "file:../../packages/core/dist",
-    "express": "^4.15.2",
-    "rxjs": "~7.6.0",
-    "tslib": "^2.4.0",
-    "zone.js": "~0.13.0"
+    "@ngx-env/builder": "file:../../packages/builder/dist",
+    "express": "^4.18.2",
+    "rxjs": "~7.8.0",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.14.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^16.0.0",
-    "@angular/cli": "^16.0.0",
-    "@angular/compiler-cli": "^16.0.0",
-    "@nguniversal/builders": "^16.0.0",
-    "@ngx-env/builder": "file:../../packages/builder",
-    "@types/express": "^4.17.0",
-    "@types/jasmine": "~3.8.0",
-    "@types/node": "^12.11.1",
-    "cross-env": "^7.0.3",
-    "jasmine-core": "~3.8.0",
-    "karma": "~6.3.0",
-    "karma-chrome-launcher": "~3.0.0",
-    "karma-coverage": "~2.0.3",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "~1.7.0",
-    "typescript": "~4.9.3"
+    "@angular-devkit/build-angular": "^17.0.0",
+    "@angular/cli": "^17.0.0",
+    "@angular/compiler-cli": "^17.0.0",
+    "@types/express": "^4.17.17",
+    "@types/jasmine": "~5.1.0",
+    "@types/node": "^18.18.0",
+    "jasmine-core": "~5.1.0",
+    "karma": "~6.4.0",
+    "karma-chrome-launcher": "~3.2.0",
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "~2.1.0",
+    "typescript": "~5.2.2"
   }
 }
