--- apps/ngx-env-demo/src/app/app.module.ts
@@ -1,11 +1,0 @@
-import { NgModule } from '@angular/core';
-import { BrowserModule } from '@angular/platform-browser';
-import { NgxEnvModule } from '@ngx-env/core';
-import { AppComponent } from './app.component';
-
-@NgModule({
-  declarations: [AppComponent],
-  imports: [BrowserModule.withServerTransition({ appId: 'serverApp' }), NgxEnvModule],
-  bootstrap: [AppComponent],
-})
-export class AppModule { }
