+++ apps/ngx-env-demo/src/env.d.ts
