+++ apps/ngx-env-demo/src/index.html
@@ -5,7 +5,7 @@
     <base href="/" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <link rel="icon" type="image/x-icon" href="favicon.ico" />
-    <title>%NGX_VERSION%</title>
+    <title>%NG_APP_VERSION%</title>
   </head>
   <body>
     <h1>%NG_APP_ENV%</h1>
