+++ apps/ngx-env-demo/tsconfig.app.json
@@ -3,11 +3,14 @@
   "extends": "./tsconfig.json",
   "compilerOptions": {
     "outDir": "./out-tsc/app",
-    "types": []
+    "types": [
+      "node"
+    ]
   },
   "files": [
     "src/main.ts",
-    "src/polyfills.ts"
+    "src/main.server.ts",
+    "server.ts",
   ],
   "include": [
     "src/**/*.d.ts"
