--- apps/ngx-env-demo/tsconfig.server.json
@@ -1,19 +1,0 @@
-/* To learn more about this file see: https://angular.io/config/tsconfig. */
-{
-  "extends": "./tsconfig.app.json",
-  "compilerOptions": {
-    "outDir": "./out-tsc/server",
-    "types": [
-      "node",
-      "@angular/localize"
-    ]
-  },
-  "files": [
-    "src/main.server.ts",
-    "server.ts",
-    "env.d.ts"
-  ],
-  "exclude": [
-    "src/env.d.ts"
-  ]
-}
