+++ apps/ngx-env-demo/tsconfig.spec.json
@@ -7,10 +7,6 @@
       "jasmine"
     ]
   },
-  "files": [
-    "src/test.ts",
-    "src/polyfills.ts"
-  ],
   "include": [
     "src/**/*.spec.ts",
     "src/**/*.d.ts"
