--- apps/ngx-env-v17-demo/.env
@@ -1,3 +1,0 @@
-USER_HOME=/home/chihab
-NG_APP_VERSION=IGNORED_SINCE_SET_IN_ENV_LOCAL
-NG_APP_USER_HOME=$USER_HOME
