--- apps/ngx-env-v17-demo/.env.local
@@ -1,1 +1,0 @@
-NG_APP_VERSION=17
