--- apps/ngx-env-v17-demo/src/app/app.component.html
@@ -1,28 +1,0 @@
-<table>
-  <thead>
-    <tr>
-      <th>NAME</th>
-      <th>VALUE</th>
-    </tr>
-  </thead>
-  <tbody>
-    <tr>
-      <td>ENV</td>
-      <td>
-        <span id="ENV">{{ env }}</span>
-      </td>
-    </tr>
-    <tr>
-      <td>VERSION</td>
-      <td>
-        <span id="VERSION"> {{ version }} </span>
-      </td>
-    </tr>
-    <tr>
-      <td>USER_HOME</td>
-      <td>
-        <span id="USER_HOME">{{ home }}</span>
-      </td>
-    </tr>
-  </tbody>
-</table>
