--- apps/ngx-env-v17-demo/src/environments/environment.prod.ts
@@ -1,6 +1,0 @@
-export const environment = {
-  production: true,
-  env: {
-    NGX_VERSION: import.meta.env['NGX_VERSION'],
-  },
-};
