--- apps/ngx-env-v17-demo/src/index.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <base href="/" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <link rel="icon" type="image/x-icon" href="favicon.ico" />
-    <title>%NG_APP_VERSION%</title>
-  </head>
-  <body>
-    <h1>%NG_APP_ENV%</h1>
-    <app-root></app-root>
-  </body>
-</html>
