--- apps/ngx-env-v17-demo/tsconfig.app.json
@@ -1,18 +1,0 @@
-/* To learn more about this file see: https://angular.io/config/tsconfig. */
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "outDir": "./out-tsc/app",
-    "types": [
-      "node"
-    ]
-  },
-  "files": [
-    "src/main.ts",
-    "src/main.server.ts",
-    "server.ts",
-  ],
-  "include": [
-    "src/**/*.d.ts"
-  ]
-}
